Fix parts package route export
